restructure sciencs and technology - new selection list for apparent temperature related tools
