+++ lib/widgets/selector_lists/apparent_temperature.dart
@@ -1,0 +1,32 @@
+import 'package:flutter/material.dart';
+import 'package:gc_wizard/widgets/common/gcw_tool.dart';
+import 'package:gc_wizard/widgets/common/gcw_toollist.dart';
+import 'package:gc_wizard/widgets/registry.dart';
+import 'package:gc_wizard/widgets/selector_lists/gcw_selection.dart';
+import 'package:gc_wizard/widgets/tools/science_and_technology/apparent_temperature/heat_index.dart';
+import 'package:gc_wizard/widgets/tools/science_and_technology/apparent_temperature/humidex.dart';
+import 'package:gc_wizard/widgets/tools/science_and_technology/apparent_temperature/summer_simmer.dart';
+import 'package:gc_wizard/widgets/tools/science_and_technology/apparent_temperature/windchill.dart';
+import 'package:gc_wizard/widgets/utils/common_widget_utils.dart';
+
+class ApparentTemperatureSelection extends GCWSelection {
+  @override
+  Widget build(BuildContext context) {
+
+    final List<GCWTool> _toolList =
+    Registry.toolList.where((element) {
+      return [
+        className(HeatIndex()),
+        className(Humidex()),
+        className(SummerSimmerIndex()),
+        className(Windchill())
+      ].contains(className(element.tool));
+    }).toList();
+
+    return Container(
+        child: GCWToolList(
+            toolList: _toolList
+        )
+    );
+  }
+}
