+++ lib/widgets/tools/science_and_technology/apparent_temperature/summer_simmer.dart
